move trainModal component to bh
