+++ src/components/repository/RepositoryViewBase/RepositoryViewBase.vue
@@ -174,7 +174,7 @@ import { mapActions, mapGetters } from 'vuex';
 import Layout from '@/components/shared/Layout';
 import RepositoryInfo from '@/components/repository/RepositoryInfo';
 import RepositoryNavigation from './RepositoryNavigation';
-import TrainModal from '@/components-v1/repository/TrainModal';
+import TrainModal from '@/components/repository/TrainModal';
 import TrainResponse from '@/components-v1/repository/TrainResponse';
 import RequestAuthorizationForm from '@/components-v1/repository/RequestAuthorizationForm';
